+++ tests/bootstrap.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * PHP library for obtain headers MIME.
+ * 
+ * @author     Josantonius - info@josantonius.com
+ * @copyright  Copyright (c) 2016 - 2017
+ * @license    https://opensource.org/licenses/MIT - The MIT License (MIT)
+ * @link       https://github.com/Josantonius/PHP-MimeType
+ * @since      1.1.3
+ */
+
+$loader = require __DIR__ . '/../src/bootstrap.php';
+
+$loader->add('Josantonius\MimeType\Tests', __DIR__);