+++ CHANGELOG.md
@@ -1,20 +1,48 @@
 # CHANGELOG
 
+## 1.1.3 - 2017-09-10
+
+* Unit tests supported by `PHPUnit` were added.
+
+* Deleted `Josantonius\MimeType\MimeType::load()` method.
+* Deleted `Josantonius\MimeType\MimeType::getAll()` method.
+
+* Added `Josantonius\MimeType\MimeType::get()` method.
+
+* Deleted `Josantonius\MimeType\Tests\MimeTypeTest::testGetAll()` method.
+
+* Added `Josantonius\MimeType\MimeType::get()` method.
+
+* Added `Josantonius\MimeType\Tests\MimeTypeTest::_getFromFile()` method.
+
+* Added `MimeType/src/bootstrap.php` file
+
+* Added `MimeType/tests/bootstrap.php` file.
+
+* Added `MimeType/phpunit.xml.dist` file.
+* Added `MimeType/_config.yml` file.
+* Added `MimeType/.travis.yml` file.
+
 ## 1.1.2 - 2017-07-16
+
 * Deleted `Josantonius\MimeType\Exception\MimeTypeException` class.
 * Deleted `Josantonius\MimeType\Exception\Exceptions` abstract class.
 * Deleted `Josantonius\MimeType\Exception\MimeTypeException->__construct()` method.
 
 ## 1.1.1 - 2017-03-18
+
 * Some files were excluded from download and comments and readme files were updated.
 
 ## 1.1.0 - 2017-01-31
+
 * Compatible with PHP 5.6 or higher.
 
 ## 1.0.0 - 2017-01-31
+
 * Compatible only with PHP 7.0 or higher. In the next versions, the library will be modified to make it compatible with PHP 5.6 or higher.
 
 ## 1.0.0 - 2016-12-19
+
 * Added `Josantonius\MimeType\MimeType` class.
 * Added `Josantonius\MimeType\MimeType::load()` method.
 * Added `Josantonius\MimeType\MimeType::getMimeFromExtension()` method.
@@ -22,11 +50,13 @@
 * Added `Josantonius\MimeType\MimeType::getAll()` method.
 
 ## 1.0.0 - 2016-12-19
+
 * Added `Josantonius\MimeType\Exception\MimeTypeException` class.
 * Added `Josantonius\MimeType\Exception\Exceptions` abstract class.
 * Added `Josantonius\MimeType\Exception\MimeTypeException->__construct()` method.
 
 ## 1.0.0 - 2016-12-19
+
 * Added `Josantonius\MimeType\Tests\MimeTypeTest` class.
 * Added `Josantonius\MimeType\Tests\MimeTypeTest::testGetMimeFromExtension()` method.
 * Added `Josantonius\MimeType\Tests\MimeTypeTest::testGetMimeFromExtensionUndefined()` method.